Extended the number of counts and added DECIMATE option
This allows to operate motors with large reductions and still keep track of a maximum number of steps while. DECIMATE allows to divide the total count to create an approximate position representation that can be displayed on the  LCD screen.

+++ Software/tunerControllerGen1/src/hwControl/StepperMotor.cpp
@@ -6,7 +6,7 @@ void StepperMotor::initMotorState() {
     // digitalWrite(DIRECTION_PIN, HIGH);
 }
 
-int StepperMotor::getRotationCount() {
+long StepperMotor::getRotationCount() {
     return rotationCount;
 }
 
@@ -53,7 +53,7 @@ StepperMotor *StepperMotor::getInstance() {
 }
 
 void StepperMotor::setRotationCount(int rotCount) {
-    rotationCount = rotCount;
+    rotationCount = (long)rotCount;
 }
 
 StepperMotor::StepperMotor() {}
